ai_research_agent/requirements.txt
crewai
langchain
openai
serpapi
python-dotenv
beautifulsoup4
requests
